feat - Add Lcd lib

+++ MPPT_Solar_Charge_Controller/Application/Analog.h
@@ -2,7 +2,7 @@
  * Analog.h
  *
  *  Created on: May 14, 2024
- *      Author: This PC
+ *      Author: binhhv.23.1.99@gmail.com
  */
 
 #ifndef ANALOG_H_
@@ -29,17 +29,17 @@ namespace blib
             virtual ~Analog();
             void readAnalog();
             void sampling();
-            double getVin() const;
-            double getIin() const;
-            double getVout() const;
-            double getIout() const;
-            double getTemp() const;
+            float getVin() const;
+            float getIin() const;
+            float getVout() const;
+            float getIout() const;
+            float getTemp() const;
         private:
-            double convertAdcChannel1(uint32_t adcValue);    // Convert adcValue -> Vin
-            double convertAdcChannel2(uint32_t adcValue);    // Convert adcValue -> Iin
-            double convertAdcChannel3(uint32_t adcValue);    // Convert adcValue -> Vout
-            double convertAdcChannel4(uint32_t adcValue);    // Convert adcValue -> Iout
-            double convertAdcChannel5(uint32_t adcValue);    // Convert adcValue -> Temperature
+            float convertAdcChannel1(uint32_t adcValue);    // Convert adcValue -> Vin
+            float convertAdcChannel2(uint32_t adcValue);    // Convert adcValue -> Iin
+            float convertAdcChannel3(uint32_t adcValue);    // Convert adcValue -> Vout
+            float convertAdcChannel4(uint32_t adcValue);    // Convert adcValue -> Iout
+            float convertAdcChannel5(uint32_t adcValue);    // Convert adcValue -> Temperature
         private:
             // Channel 1: VIN
             // Channel 2: IIN
@@ -47,23 +47,23 @@ namespace blib
             // Channel 4: IOUT
             // Channel 5: TEMP
             static const uint8_t NUMBER_OF_CHANNELS = 5U;
-            static constexpr double CHANNEL_1_RATIO = 10.0;
-            static constexpr double CHANNEL_2_RATIO = 3.0;
-            static constexpr double CHANNEL_3_RATIO = 5.0;
-            static constexpr double CHANNEL_4_RATIO = 10.0;
-            static constexpr double CHANNEL_5_RATIO = 10.0;
+            static constexpr float CHANNEL_1_RATIO = 10.0f;
+            static constexpr float CHANNEL_2_RATIO = 3.0f;
+            static constexpr float CHANNEL_3_RATIO = 5.0f;
+            static constexpr float CHANNEL_4_RATIO = 10.0f;
+            static constexpr float CHANNEL_5_RATIO = 10.0f;
 
-            static constexpr double ACS_SENSITIVITY = 0.1;
+            static constexpr float ACS_SENSITIVITY = 0.1f;
 
             uint32_t mAdcValues[NUMBER_OF_CHANNELS] = { 0 };
             ADC_HandleTypeDef *mHadc;
             DMA_HandleTypeDef *mHdmaAdc;
 
-            double mVin;
-            double mIin;
-            double mVout;
-            double mIout;
-            double mTemp;
+            float mVin;
+            float mIin;
+            float mVout;
+            float mIout;
+            float mTemp;
     };
 }
